backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
